feat: add InternalMoveLineDetailsPicker
