+++ packages/apps/stock/src/components/templates/index.js
@@ -30,6 +30,7 @@ export {default as InternalMoveGeneralFixedItems} from './InternalMoveGeneralFix
 export {default as InternalMoveLineCard} from './InternalMoveLineCard/InternalMoveLineCard';
 export {default as InternalMoveLineDetailsFixedItems} from './InternalMoveLineDetailsFixedItems/InternalMoveLineDetailsFixedItems';
 export {default as InternalMoveLineDetailsQuantityCard} from './InternalMoveLineDetailsQuantityCard/InternalMoveLineDetailsQuantityCard';
+export {default as InternalMoveLineDetailsPicker} from './InternalMoveLineDetailsPicker/InternalMoveLineDetailsPicker';
 export {default as InternalMoveLineNotes} from './InternalMoveLineNotes/InternalMoveLineNotes';
 export {default as InventoryCard} from './InventoryCard/InventoryCard';
 export {default as InventoryHeader} from './InventoryHeader/InventoryHeader';
